move the json model
